requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
google-generativeai>=0.3.0